requirements.txt
django
djangorestframework
django-extensions
django-environ
psycopg2-binary
pydantic
django-redis
drf-spectacular
django-cors-headers
djangorestframework-simplejwt>=5.2.0
PyJWT>=2.6.0
